djdjjdjdjd

+++ src/main/src/main/java/com/gerry/jnshu/response/ResultCode.java
@@ -4,15 +4,15 @@ public enum ResultCode {
     SUCCESS(200, "操作成功"),
     FAILED(500, "操作失败"),
     VALIDATE_FAILED(404, "参数检验失败");
-    private long code;
+    private int code;
     private String message;
 
-    private ResultCode(long code, String message) {
+    private ResultCode(int code, String message) {
         this.code = code;
         this.message = message;
     }
 
-    public long getCode() {
+    public int getCode() {
         return code;
     }
